+++ src/main/java/kitchenpos/application/MenuGroupService.java
@@ -1,7 +1,7 @@
 package kitchenpos.application;
 
 import java.util.List;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menu.MenuGroup;
 import kitchenpos.dto.request.MenuGroupCreateRequest;
 import kitchenpos.repository.MenuGroupRepository;
 import org.springframework.stereotype.Service;
