+++ src/main/java/kitchenpos/application/MenuService.java
@@ -3,10 +3,10 @@ package kitchenpos.application;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menu.MenuGroup;
+import kitchenpos.domain.menu.MenuProduct;
+import kitchenpos.domain.menu.Product;
 import kitchenpos.dto.request.MenuCreateRequest;
 import kitchenpos.dto.request.MenuProductCreateRequest;
 import kitchenpos.repository.MenuGroupRepository;
@@ -36,6 +36,7 @@ public class MenuService {
         final Menu menu = new Menu(
                 request.getName(),
                 request.getPrice(),
+                /// TODO: 2023/10/25  menuGroupId만 받고 Validator 사용
                 findMenuGroup(request.getMenuGroupId()),
                 createMenuProducts(request.getMenuProducts())
         );
