+++ src/main/java/kitchenpos/domain/menu/Menu.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
@@ -24,6 +24,7 @@ public class Menu {
     private MenuName name;
     @Embedded
     private Price price;
+    // TODO 불필요한 의존성 끊기
     @ManyToOne(optional = false)
     @JoinColumn(name = "menu_group_id", foreignKey = @ForeignKey(name = "fk_menu_to_menu_group"))
     private MenuGroup menuGroup;
