+++ src/main/java/kitchenpos/domain/menu/MenuProducts.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
 import java.util.ArrayList;
 import java.util.List;
