+++ src/main/java/kitchenpos/domain/order/OrderLineItems.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
 import java.util.ArrayList;
 import java.util.List;
