+++ src/main/java/kitchenpos/domain/order/TableGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -30,16 +30,16 @@ public class TableGroup {
     public TableGroup() {
     }
 
-    private void validateOrderTablesSize(final List<OrderTable> orderTables) {
-        if (CollectionUtils.isEmpty(orderTables) || orderTables.size() < 2) {
-            throw new IllegalArgumentException("단체에 속한 테이블은 최소 2개 이상이어야 합니다.");
-        }
-    }
-
     public void addOrderTables(final List<OrderTable> orderTables) {
         validateOrderTablesSize(orderTables);
         orderTables.forEach(orderTable -> orderTable.group(this));
         this.orderTables = orderTables;
+    }
+
+    private void validateOrderTablesSize(final List<OrderTable> orderTables) {
+        if (CollectionUtils.isEmpty(orderTables) || orderTables.size() < 2) {
+            throw new IllegalArgumentException("단체에 속한 테이블은 최소 2개 이상이어야 합니다.");
+        }
     }
 
     public Long getId() {
