+++ src/main/java/kitchenpos/dto/response/OrderResponse.java
@@ -3,8 +3,8 @@ package kitchenpos.dto.response;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.vo.OrderStatus;
 
 public class OrderResponse {
