+++ src/main/java/kitchenpos/dto/response/OrderTableResponse.java
@@ -3,8 +3,8 @@ package kitchenpos.dto.response;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.order.TableGroup;
 
 public class OrderTableResponse {
