+++ src/main/java/kitchenpos/dto/response/ProductResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.dto.response;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.menu.Product;
 
 public class ProductResponse {
 
@@ -32,7 +32,7 @@ public class ProductResponse {
                 .map(ProductResponse::from)
                 .collect(Collectors.toList());
     }
-    
+
     public long getId() {
         return id;
     }
