+++ src/main/java/kitchenpos/dto/response/TableGroupResponse.java
@@ -2,7 +2,7 @@ package kitchenpos.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.order.TableGroup;
 
 public class TableGroupResponse {
