+++ src/main/java/kitchenpos/repository/MenuRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.repository;
 
 import java.util.List;
-import kitchenpos.domain.Menu;
+import kitchenpos.domain.menu.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
