+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -9,9 +9,9 @@ import java.math.BigDecimal;
 import java.util.List;
 import java.util.Objects;
 import kitchenpos.IntegrationTest;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menu.MenuGroup;
+import kitchenpos.domain.menu.Product;
 import kitchenpos.dto.request.MenuCreateRequest;
 import kitchenpos.dto.request.MenuGroupCreateRequest;
 import kitchenpos.fixture.RequestParser;
