+++ src/test/java/kitchenpos/domain/menu/MenuGroupTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
