+++ src/test/java/kitchenpos/fixture/ProductFixture.java
@@ -1,7 +1,7 @@
 package kitchenpos.fixture;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.menu.Product;
 
 public class ProductFixture {
